requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cs50==9.4.0
dotenv==0.9.9
exceptiongroup==1.3.0
Flask==3.1.1
Flask-Session==0.8.0
google-auth==2.40.3
google-genai==1.26.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tenacity==8.5.0
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
